requirements.txt
networkx
mysql-connector
gitpython
python-bugzilla
pygithub
selenium
py-stackexchange
beautifulsoup4
slacker